NodeApi becomes a module, add NodeApi::Base class
